Updaating some spec references
